Make Python aware of required arguments
By removing the default value `None` from both required arguments of
the method `discharge`, Python is aware that these two arguments are
not optional and will throw an error message if one of them is missing.
We can therefore remove the if-statements checking for `None`.

I also added a description to the argument `quiet`.