Move all imports to the beginning of the file
It is common in Python to put all import-statements at the beginning of
the code file.  The script `discharge.py` contained an import within
the function `p2mp` within the method `outlets`; this import is now
moved to the top.

Having moved the import to the top, the function `p2mp` served merely as
a synonym for `unary_union`, so for simplicity, both calls to `p2mp` were
replaced with `unary_union` and `p2mp` was removed.